chore: move some service tests to a dedicated file
